float to long double

+++ tasks/search/search.cpp
@@ -91,16 +91,16 @@ std::vector<std::string_view> Search(std::string_view text, std::string_view que
     std::vector<std::string_view> strings_of_text = {};
     std::vector<size_t> strings_amounts_of_words = {};
     std::vector<std::unordered_map<std::string_view, size_t>> strings_maps = {};
-    std::vector<std::unordered_map<std::string_view, float>> self_tfs = {};
-    std::vector<std::unordered_map<std::string_view, float>> self_idfs = {};
-    std::vector<std::pair<float, size_t>> tf_idfs = {};
+    std::vector<std::unordered_map<std::string_view, long double>> self_tfs = {};
+    std::vector<std::unordered_map<std::string_view, long double>> self_idfs = {};
+    std::vector<std::pair<long double, size_t>> tf_idfs = {};
     SeparateTextToStrings(text, strings_of_text);
     for (size_t i = 0; i < strings_of_text.size(); ++i) {
         StringOfTextToMap(strings_of_text[i], strings_amounts_of_words, strings_maps);
-        std::unordered_map<std::string_view, float> self_tf = {};
-        std::unordered_map<std::string_view, float> self_idf = {};
+        std::unordered_map<std::string_view, long double> self_tf = {};
+        std::unordered_map<std::string_view, long double> self_idf = {};
         for (auto query_word : query_words) {
-            self_tf[query_word] = static_cast<float>(strings_maps[i][query_word]) / strings_amounts_of_words[i];
+            self_tf[query_word] = static_cast<long double>(strings_maps[i][query_word]) / strings_amounts_of_words[i];
             if (self_idf.contains(query_word)) {
                 self_idf[query_word] += 1;
             } else {
@@ -112,10 +112,10 @@ std::vector<std::string_view> Search(std::string_view text, std::string_view que
     }
 
     for (size_t i = 0; i < strings_of_text.size(); ++i) {
-        float tf_idf = 0;
+        long double tf_idf = 0;
         for (auto query_word : query_words) {
             tf_idf += self_tfs[i][query_word] *
-                      std::log(static_cast<float>(strings_of_text.size()) / self_idfs[i][query_word]);
+                      std::log(static_cast<long double>(strings_of_text.size()) / self_idfs[i][query_word]);
         }
         tf_idfs.emplace_back(std::pair(tf_idf, strings_of_text.size() - i));
     }
